+++ src/OpenLoco/src/Ui/Windows/EditKeyboardShortcut.cpp
@@ -1,6 +1,4 @@
-#include "Graphics/Colour.h"
 #include "Graphics/ImageIds.h"
-#include "Graphics/TextRenderer.h"
 #include "Input/Shortcuts.h"
 #include "Localisation/FormatArguments.hpp"
 #include "Localisation/StringIds.h"
@@ -10,6 +8,7 @@
 #include "Ui/Widgets/CaptionWidget.h"
 #include "Ui/Widgets/FrameWidget.h"
 #include "Ui/Widgets/ImageButtonWidget.h"
+#include "Ui/Widgets/LabelWidget.h"
 #include "Ui/Widgets/PanelWidget.h"
 #include "Ui/WindowManager.h"
 #include <OpenLoco/Engine/Input/ShortcutManager.h>
@@ -19,15 +18,19 @@ using namespace OpenLoco::Input;
 
 namespace OpenLoco::Ui::Windows::EditKeyboardShortcut
 {
-    static constexpr Ui::Size kWindowSize = { 280, 72 };
+    static constexpr Ui::Size kWindowSize = { 280, 78 };
 
     static uint8_t _editingShortcutIndex;
+    static KeyModifier _pressedModifiers;
 
     static constexpr auto _widgets = makeWidgets(
         Widgets::Frame({ 0, 0 }, kWindowSize, WindowColour::primary),
         Widgets::Caption({ 1, 1 }, { kWindowSize.width - 2, 13 }, Widgets::Caption::Style::whiteText, WindowColour::primary, StringIds::change_keyboard_shortcut),
         Widgets::ImageButton({ 265, 2 }, { 13, 13 }, WindowColour::primary, ImageIds::close_button, StringIds::tooltip_close_window),
-        Widgets::Panel({ 0, 15 }, { kWindowSize.width, 57 }, WindowColour::secondary));
+        Widgets::Panel({ 0, 15 }, { kWindowSize.width, kWindowSize.height - 15 }, WindowColour::secondary),
+        Widgets::Label({ 4, 20 }, { kWindowSize.width - 8, 12 }, WindowColour::secondary, ContentAlign::center, StringIds::change_keyboard_shortcut_desc),
+        Widgets::Label({ 4, 34 }, { kWindowSize.width - 8, 12 }, WindowColour::secondary, ContentAlign::center, StringIds::black_quoted_stringid),
+        Widgets::Label({ 4, 54 }, { kWindowSize.width - 8, 12 }, WindowColour::secondary, ContentAlign::center, StringIds::black_stringid));
 
     static const WindowEventList& getEvents();
 
@@ -39,6 +42,9 @@ namespace OpenLoco::Ui::Windows::EditKeyboardShortcut
             caption,
             close,
             panel,
+            description,
+            shortcutName,
+            pressedKeys,
         };
     }
 
@@ -47,6 +53,7 @@ namespace OpenLoco::Ui::Windows::EditKeyboardShortcut
     {
         WindowManager::close(WindowType::editKeyboardShortcut);
         _editingShortcutIndex = shortcutIndex;
+        _pressedModifiers = KeyModifier::none;
 
         auto window = WindowManager::createWindow(WindowType::editKeyboardShortcut, kWindowSize, WindowFlags::none, getEvents());
 
@@ -97,17 +104,40 @@ namespace OpenLoco::Ui::Windows::EditKeyboardShortcut
         WindowManager::invalidate(WindowType::keyboardShortcuts);
     }
 
+    static void onUpdate(Window& self)
+    {
+        const auto modifiers = Input::getKeyModifier() & ~KeyModifier::cheat;
+        if (modifiers == _pressedModifiers)
+        {
+            return;
+        }
+
+        _pressedModifiers = modifiers;
+        self.invalidate();
+    }
+
+    static void prepareDraw(Window& self)
+    {
+        {
+            auto args = FormatArguments(self.widgets[Widx::description].textArgs);
+            args.push(StringIds::empty);
+        }
+
+        {
+            auto args = FormatArguments(self.widgets[Widx::shortcutName].textArgs);
+            args.push(ShortcutManager::getName(static_cast<Shortcut>(_editingShortcutIndex)));
+        }
+
+        {
+            auto args = FormatArguments(self.widgets[Widx::pressedKeys].textArgs);
+            Input::Shortcuts::pushModifierStrings(args, _pressedModifiers);
+        }
+    }
+
     // 0x004BE8DF
     static void draw(Ui::Window& self, Gfx::DrawingContext& drawingCtx)
     {
-        auto tr = Gfx::TextRenderer(drawingCtx);
-
         self.draw(drawingCtx);
-
-        FormatArguments args{};
-        args.push(ShortcutManager::getName(static_cast<Shortcut>(_editingShortcutIndex)));
-        auto point = Ui::Point(self.x + 140, self.y + 32);
-        tr.drawStringCentredWrapped(point, 272, Colour::black, StringIds::change_keyboard_shortcut_desc, args);
     }
 
     // 0x004BE821
@@ -130,6 +160,8 @@ namespace OpenLoco::Ui::Windows::EditKeyboardShortcut
 
     static constexpr WindowEventList kEvents = {
         .onMouseUp = onMouseUp,
+        .onUpdate = onUpdate,
+        .prepareDraw = prepareDraw,
         .draw = draw,
         .keyUp = onKeyUp,
     };